Use fixed-sized integer data types in GPU data structures (OEI, ERI). More swaps from 1-based to 0-based indexed to avoid superfluous arithmetic (OEI, ERI). Avoid compiling unused ERI VRR kernels. Other code clean-up.

+++ src/gpu/gpu_common.h
@@ -10,6 +10,7 @@
 #if !defined(__QUICK_GPU_COMMON_H_)
 #define __QUICK_GPU_COMMON_H_
 
+#include <stdint.h>
 #include <stdio.h>
 
 #include "../octree/gpack_common.h"
@@ -229,8 +230,8 @@ enum QUICK_METHOD
 };
 
 struct ERI_entry {
-    int IJ;
-    int KL;
+    uint8_t IJ;
+    uint8_t KL;
     QUICKDouble value;
 };
